+++ changelog.md
@@ -2,6 +2,10 @@
 
 ## v0.6.11
 
+- **Fixed:**
+  - Display:
+    - Fixed horizontal scroll bar when it first appears auto scrolling right correctly this time
+
 ## v0.6.10 2018-11-21
 
 - **Fixed:**